requirements.txt
python-dotenv==1.0.0
PyGithub==2.1.1
requests==2.31.0
openai==1.12.0
google-api-python-client==2.118.0 